sdk/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-silero
livekit-plugins-cartesia
livekit-plugins-turn-detector
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
python-dotenv
livekit-plugins-sarvam
livekit-plugins-noise-cancellation
twine
# Dependencies for health_bench evaluation
openai>=1.0.0
numpy
pandas
blobfile
requests
tqdm
jinja2